model_service/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.0.0
joblib>=1.3.0
kafka-python>=2.0.0
pydantic>=2.0.0
requests>=2.31.0
